Switch from raf to error ignore
The raf approach for preventing the RO loop limit error (#1) causes a noticeable initial render delay.

"ResizeObserver loop limit exceeded" error is fired as an `error` event at `window` object, whenever resize is triggered by the RO callback.
The error is safe to ignore as it is just a warning message, telling that the "looping" callback will be skipped in the current frame, and will be called in the next one.
We need to ignore it to prevent failing Ember tests.